orders/requirements.dev.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.4
billiard==4.1.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.5
django-baton==2.8.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-oauth-toolkit==2.3.0
django-rest-passwordreset==1.3.0
django-silk==5.0.4
django-versatileimagefield==3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
drf-yasg==1.21.7
ecdsa==0.18.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
gprof2dot==2022.7.29
httplib2==0.22.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
kombu==5.3.2
model-bakery==1.15.0
oauthlib==3.2.2
packaging==23.2
Pillow==9.5.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
rsa==4.9
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
wrapt==1.15.0
